+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In development] - Unreleased
 
+### Fixed
+
+- `NoReverseMatch` error when slugs for category, board or topic that are generated
+  from the name/subject ended up being empty when only entered special chars like
+  "@#$%" as name/subject
+
 
 ## [1.4.2] - 2022-01-23